[DDW-723] Refactor to remove duplication
