+++ source/renderer/app/components/wallet/WalletSendForm.tsx
@@ -30,8 +30,8 @@ import { FORM_VALIDATION_DEBOUNCE_WAIT } from '../../config/timingConfig';
 import { TRANSACTION_MIN_ADA_VALUE } from '../../config/walletsConfig';
 import { NUMBER_FORMATS } from '../../../../common/types/number.types';
 import AssetInput from './send-form/AssetInput';
-import WalletSendAssetsConfirmationDialog from './send-form/WalletSendAssetsConfirmationDialog';
-import WalletSendConfirmationDialogContainer from '../../containers/wallet/dialogs/WalletSendConfirmationDialogContainer';
+import { WalletSendConfirmationDialogView } from '../../containers/wallet/dialogs/send-confirmation/View';
+import { WalletSendConfirmationDialogContainer } from '../../containers/wallet/dialogs/send-confirmation/Container';
 import styles from './WalletSendForm.scss';
 import Asset from '../../domains/Asset';
 import type { HwDeviceStatus } from '../../domains/Wallet';
@@ -1138,13 +1138,12 @@ class WalletSendForm extends Component<Props, State> {
           </BorderedBox>
         )}
 
-        {isDialogOpen(WalletSendAssetsConfirmationDialog) ? (
+        {isDialogOpen(WalletSendConfirmationDialogView) ? (
           <WalletSendConfirmationDialogContainer
             receiver={receiver}
             selectedAssets={this.selectedAssets}
             assetsAmounts={this.selectedAssetsAmounts}
             amount={adaAmount.toFormat(currencyMaxFractionalDigits)}
-            amountToNaturalUnits={formattedAmountToNaturalUnits}
             totalAmount={total}
             transactionFee={fees}
             hwDeviceStatus={hwDeviceStatus}
