+++ source/renderer/app/containers/wallet/WalletSendPage.tsx
@@ -6,7 +6,7 @@ import {
   MAX_INTEGER_PLACES_IN_ADA,
 } from '../../config/numbersConfig';
 import WalletSendForm from '../../components/wallet/WalletSendForm';
-import WalletSendAssetsConfirmationDialog from '../../components/wallet/send-form/WalletSendAssetsConfirmationDialog';
+import { WalletSendConfirmationDialogView } from './dialogs/send-confirmation/View';
 import WalletTokenPicker from '../../components/wallet/tokens/wallet-token-picker/WalletTokenPicker';
 import { WALLET_ASSETS_ENABLED } from '../../config/walletsConfig';
 import Asset from '../../domains/Asset';
@@ -71,7 +71,7 @@ class WalletSendPage extends Component<Props> {
     // @ts-ignore ts-migrate(2339) FIXME: Property 'isFlight' does not exist on type 'typeof... Remove this comment to see the full error message
     const { isFlight } = global;
     this.props.actions.dialogs.open.trigger({
-      dialog: WalletSendAssetsConfirmationDialog,
+      dialog: WalletSendConfirmationDialogView,
     });
 
     if (isHardwareWallet && !isFlight) {
