+++ source/renderer/app/i18n/locales/en-US.json
@@ -36,12 +36,6 @@
   "appUpdate.overlay.postponeInstall.link.label": "Postpone the update",
   "appUpdate.overlay.subtitle": "You are currently running Daedalus version <b>{currentAppVersion}</b>.<br />Daedalus version <b>{availableAppVersion}</b> is now available to download.",
   "appUpdate.overlay.title": "Software update available!",
-  "asset.transactionConfirmation.assetLabel": "Token #{assetNumber}",
-  "asset.transactionConfirmation.insufficientBalance": "Insufficient funds available. The balance of the token in this wallet is {formattedBalance} (Unformatted: {unformattedBalance})",
-  "asset.transactionConfirmation.missingToken": "The token does not exist in this wallet",
-  "asset.transactionConfirmation.unformattedAmountLabel": "unformatted amount",
-  "asset.transactionConfirmation.unformattedAmountMessageForHardwareWallets": "Native assets may specify a number of decimal places, as defined in the Cardano token registry. Daedalus uses this information to format the amount that is being sent in the transaction.<br /><br />The native token unformatted amount is the amount without these decimal places. Please ensure that you verify both amounts, as some wallet software may not yet use the Cardano token registry.",
-  "asset.transactionConfirmation.unformattedAmountMessageForSoftwareWallets": "Native assets may specify a number of decimal places, as defined in the Cardano token registry. Daedalus uses this information to format the amount that is being sent in the transaction.<br /><br />The native token unformatted amount is the amount without these decimal places. Please ensure that you verify both amounts, as some wallet software may not yet use the Cardano token registry.<br /><br />The native token unformatted amount will be displayed on the hardware wallet device during transaction confirmation.",
   "assets.assetAmount.unformattedAmount": "<em>Unformatted amount:</em> {amount}",
   "assets.assetToken.param.assetName": "Asset name",
   "assets.assetToken.param.blank": "Blank",
@@ -137,16 +131,6 @@
   "daedalus.diagnostics.dialog.systemTimeIgnored": "System time ignored",
   "daedalus.diagnostics.dialog.unknownDiskSpace": "Unknown",
   "daedalus.diagnostics.dialog.unknownDiskSpaceSupportUrl": "https://iohk.zendesk.com/hc/en-us/articles/360035654533",
-  "dapp.transaction.request.additionalData.label": "Additional data",
-  "dapp.transaction.request.error.notEnoughAda": "This wallet does not contain the minimum amount of {adaBalanceRequired} which is required for this transaction. Please select a wallet with <b>a minimum amount of {adaBalanceRequired}</b>.",
-  "dapp.transaction.request.fromWalletLabel": "From wallet",
-  "dapp.transaction.request.metaData.label": "Metadata",
-  "dapp.transaction.request.receiver.label": "Receiver",
-  "dapp.transaction.request.subtitle": "triggered from {triggeredFrom}",
-  "dapp.transaction.request.title": "Transaction request",
-  "dapp.transaction.request.transactionFee.label": "Transaction fee",
-  "dapp.transaction.request.walletsDropdown.addWalletLabel": "+ Add wallet",
-  "dapp.transaction.request.walletsDropdown.placeholder": "Select a wallet",
   "discreetMode.discreetToggle.description": "You can toggle auto discreet mode in <b>Settings</b>.",
   "discreetMode.discreetToggle.off": "Toggle discreet mode off.",
   "discreetMode.discreetToggle.on": "Toggle discreet mode on.",
@@ -1011,7 +995,6 @@
   "wallet.send.confirmationDialog.flightCandidateWarning": "<b>Warning:</b> flight candidate versions of Daedalus are connected to the Cardano mainnet. If you confirm this transaction, real ada will be transferred out of this wallet.",
   "wallet.send.confirmationDialog.passphraseFieldPlaceholder": "Type your spending password",
   "wallet.send.confirmationDialog.passphraseLabel": "Spending password",
-  "wallet.send.confirmationDialog.passwordError": "Incorrect spending password",
   "wallet.send.confirmationDialog.receiver.label": "Receiver",
   "wallet.send.confirmationDialog.submit": "Send",
   "wallet.send.confirmationDialog.title": "Confirm transaction",
