+++ source/renderer/app/i18n/locales/ja-JP.json
@@ -36,12 +36,6 @@
   "appUpdate.overlay.postponeInstall.link.label": "後で更新する",
   "appUpdate.overlay.subtitle": "このDaedalusはバージョン<b>{currentAppVersion}</b>です。<br />現在Daedalus <b>{availableAppVersion}</b>がダウンロード可能です",
   "appUpdate.overlay.title": "ソフトウェアが更新できます!",
-  "asset.transactionConfirmation.assetLabel": "Token #{assetNumber}",
-  "asset.transactionConfirmation.insufficientBalance": "資金が不足しています。このウォレットのトークン残高は{formattedBalance}です（書式なし：{unformattedBalance}）",
-  "asset.transactionConfirmation.missingToken": "このウォレットにこのトークンはありません",
-  "asset.transactionConfirmation.unformattedAmountLabel": "書式なしの金額",
-  "asset.transactionConfirmation.unformattedAmountMessageForHardwareWallets": "ブロックチェーンでは、ネイティブトークンは少数なしの整数で表現されます。<br /><br />書式化した金額は、ウォレット概要画面のネイティブトークンの設定で指定した書式で表示されています。<br /><br />トランザクションを送信する前に、小数点以下の桁数をもう一度確認してください。<br /><br />ネイティブトークンの書式なしの金額は、トランザクションの確認中にハードウェアウォレットデバイスに表示されます。",
-  "asset.transactionConfirmation.unformattedAmountMessageForSoftwareWallets": "ブロックチェーンでは、ネイティブトークンは少数なしの整数で表現されます。<br /><br />書式化した金額は、ウォレット概要画面のネイティブトークンの設定で指定した書式で表示されています。<br /><br />トランザクションを送信する前に、小数点以下の桁数をもう一度確認してください。",
   "assets.assetAmount.unformattedAmount": "<em>書式なしの金額：</em>{amount}",
   "assets.assetToken.param.assetName": "アセット名",
   "assets.assetToken.param.blank": "なし",
@@ -137,16 +131,6 @@
   "daedalus.diagnostics.dialog.systemTimeIgnored": "システム時刻無視",
   "daedalus.diagnostics.dialog.unknownDiskSpace": "不明",
   "daedalus.diagnostics.dialog.unknownDiskSpaceSupportUrl": "https://iohk.zendesk.com/hc/ja/articles/360035654533",
-  "dapp.transaction.request.additionalData.label": "追加データ",
-  "dapp.transaction.request.error.notEnoughAda": "このウォレットにはこのトランザクションに必要な最低額{adaBalanceRequired}が入っていません。<b>最低額{adaBalanceRequired}</b>の入ったウォレットを選択してください。",
-  "dapp.transaction.request.fromWalletLabel": "送金元ウォレット",
-  "dapp.transaction.request.metaData.label": "メタデータ",
-  "dapp.transaction.request.receiver.label": "送金先",
-  "dapp.transaction.request.subtitle": "トリガー：{triggeredFrom}",
-  "dapp.transaction.request.title": "トランザクションリクエスト",
-  "dapp.transaction.request.transactionFee.label": "トランザクション手数料",
-  "dapp.transaction.request.walletsDropdown.addWalletLabel": "+ ウォレット追加",
-  "dapp.transaction.request.walletsDropdown.placeholder": "ウォレットを選択してください",
   "discreetMode.discreetToggle.description": "<b>設定</b>で自動ディスクリートモードに切り替えられます。",
   "discreetMode.discreetToggle.off": "ディスクリートモードオフ.",
   "discreetMode.discreetToggle.on": "ディスクリートモードオン.",
@@ -1011,7 +995,6 @@
   "wallet.send.confirmationDialog.flightCandidateWarning": "<b>警告：</b>Daedalus Flight候補版はCardanoメインネットに接続されています。このトランザクションを確認すると、実際のADAがこのウォレットから送金されます。",
   "wallet.send.confirmationDialog.passphraseFieldPlaceholder": "送金時パスワードを入力して下さい",
   "wallet.send.confirmationDialog.passphraseLabel": "送金時パスワード",
-  "wallet.send.confirmationDialog.passwordError": "送金時パスワードが間違っています",
   "wallet.send.confirmationDialog.receiver.label": "送金先",
   "wallet.send.confirmationDialog.submit": "送金する",
   "wallet.send.confirmationDialog.title": "トランザクションの確認",
