+++ storybook/stories/wallets/send/WalletSend.stories.tsx
@@ -17,10 +17,9 @@ import Wallet, {
 } from '../../../../source/renderer/app/domains/Wallet';
 // Screens
 import WalletSendForm from '../../../../source/renderer/app/components/wallet/WalletSendForm';
-import WalletSendAssetsConfirmationDialog from '../../../../source/renderer/app/components/wallet/send-form/WalletSendAssetsConfirmationDialog';
-import WalletSendConfirmationDialog from '../../../../source/renderer/app/components/wallet/send-form/WalletSendConfirmationDialog';
 import { formattedAmountToNaturalUnits } from '../../../../source/renderer/app/utils/formatters';
 import type { WalletTokens } from '../../../../source/renderer/app/api/assets/types';
+import { WalletSendConfirmationDialogView } from '../../../../source/renderer/app/containers/wallet/dialogs/send-confirmation/View';
 
 const allAssets = [
   // @ts-ignore ts-migrate(2554) FIXME: Expected 7 arguments, but got 5.
@@ -376,32 +375,28 @@ storiesOf('Wallets|Send', module)
     };
     return (
       <div>
-        <WalletSendAssetsConfirmationDialog
+        <WalletSendConfirmationDialogView
           amount="20.000000"
+          areTermsAccepted={boolean('areTermsAccepted', true)}
           wallet={wallet}
           totalAmount={new BigNumber('21.000000')}
-          // @ts-ignore ts-migrate(2769) FIXME: No overload matches this call.
-          sender={
-            generateWallet('Wallet name', '45119903750165', walletTokens).id
-          }
           receiver={generateHash()}
           selectedAssets={confirmationTokens}
-          allAvailableTokens={confirmationTokens}
+          assetTokens={confirmationTokens}
           assetsAmounts={confirmationTokensAmounts}
           transactionFee="1.000000"
-          amountToNaturalUnits={formattedAmountToNaturalUnits}
-          onSubmit={() => null}
-          isSubmitting={false}
-          error={null}
-          isFlight={false}
-          onCancel={() => null}
-          onExternalLinkClick={() => null}
           hwDeviceStatus={HwDeviceStatuses.CONNECTING}
+          isFlight={boolean('isFlight', false)}
           isHardwareWallet={boolean('isHardwareWallet', false)}
-          onInitiateTransaction={() => null}
-          onCopyAssetParam={() => {}}
+          isSubmitting={boolean('isSubmitting', false)}
           isTrezor={boolean('isTrezor', false)}
           formattedTotalAmount="21.000000"
+          error={null}
+          onCancel={() => null}
+          onSubmitCb={() => null}
+          onTermsCheckboxClick={() => null}
+          onCopyAssetParam={() => null}
+          onExternalLinkClick={() => null}
         />
       </div>
     );
@@ -415,31 +410,28 @@ storiesOf('Wallets|Send', module)
     };
     return (
       <div>
-        <WalletSendConfirmationDialog
+        <WalletSendConfirmationDialogView
+          wallet={wallet}
+          receiver={generateHash()}
           amount="20.000000"
           totalAmount={new BigNumber('21.000000')}
-          wallet={wallet}
-          // @ts-ignore ts-migrate(2769) FIXME: No overload matches this call.
-          sender={
-            generateWallet('Wallet name', '45119903750165', walletTokens).id
-          }
-          receiver={generateHash()}
           transactionFee="1.000000"
-          amountToNaturalUnits={formattedAmountToNaturalUnits}
-          onSubmit={() => null}
-          isSubmitting={false}
-          error={null}
-          isFlight={false}
-          onCancel={() => null}
-          onExternalLinkClick={() => null}
+          selectedAssets={[]}
+          assetTokens={confirmationTokens}
+          assetsAmounts={confirmationTokensAmounts}
           hwDeviceStatus={HwDeviceStatuses.CONNECTING}
-          isHardwareWallet={boolean('isHardwareWallet', false)}
-          onInitiateTransaction={() => null}
-          walletName={
-            generateWallet('TrueUSD', '15119903750165', walletTokens).name
-          }
+          areTermsAccepted={boolean('areTermsAccepted', true)}
+          isFlight={boolean('isFlight', false)}
           isTrezor={boolean('isTrezor', false)}
+          isSubmitting={boolean('isSubmitting', false)}
+          isHardwareWallet={boolean('isHardwareWallet', false)}
           formattedTotalAmount="21.000000"
+          error={null}
+          onCancel={() => null}
+          onSubmitCb={() => null}
+          onTermsCheckboxClick={() => null}
+          onCopyAssetParam={() => null}
+          onExternalLinkClick={() => null}
         />
       </div>
     );
